% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{get_dynamics_table}
\alias{get_dynamics_table}
\title{Get Dynamics365 Table}
\usage{
get_dynamics_table(table_id)
}
\arguments{
\item{table_id}{character table name}
}
\value{

}
\description{
Get Dynamics365 Table
}
